finish role and permission feature

+++ app/Http/Controllers/DepartmentController.php
@@ -12,10 +12,18 @@ use Yajra\DataTables\Facades\DataTables;
 class DepartmentController extends Controller
 {
     public function index () {
+        if (!auth()->user()->can('view_department')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('department.index');
     }
 
     public function ssd () {
+        if (!auth()->user()->can('view_department')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $data = Department::query();
         return Datatables::of($data)
             ->editColumn('updated_at', function($each) {
@@ -25,8 +33,17 @@ class DepartmentController extends Controller
                 return null;
             })
             ->addColumn('action', function($each) {
-                $edit_icon = '<a href="'.route('department.edit', $each->id).'" class="text-warning"><i class="fas fa-edit"></i></a>';
-                $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="'.$each->id.'"><i class="fas fa-trash-alt"></i></a>';
+                $edit_icon = '';
+                $delete_icon = '';
+
+                if (auth()->user()->can('edit_department')) {
+                    $edit_icon = '<a href="'.route('department.edit', $each->id).'" class="text-warning"><i class="fas fa-edit"></i></a>';
+                }
+
+                if (auth()->user()->can('delete_department')) {
+                    $delete_icon = '<a href="#" class="text-danger delete-btn" data-id="'.$each->id.'"><i class="fas fa-trash-alt"></i></a>';
+                }
+
 
                 return '<div class="action-icon text-start">'.$edit_icon.$delete_icon.'</div>';
             })
@@ -35,11 +52,17 @@ class DepartmentController extends Controller
     }
 
     public function create () {
+        if (!auth()->user()->can('create_department')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('department.create');
     }
 
     public function store (StoreDepartment $request) {
-
+        if (!auth()->user()->can('create_department')) {
+            abort(403, 'Unauthorized Action');
+        }
         $department = new Department();
         $department->title = $request->title;
         $department->save();
@@ -48,11 +71,19 @@ class DepartmentController extends Controller
     }
 
     public function edit ($id) {
+        if (!auth()->user()->can('edit_department')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $department = Department::findOrFail($id);
         return view('department.edit', compact('department'));
     }
 
     public function update($id, UpdateDepartment $request) {
+        if (!auth()->user()->can('edit_department')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $department = Department::findOrFail($id);
 
         $department->title = $request->title;
@@ -64,6 +95,10 @@ class DepartmentController extends Controller
     
 
     public function destroy ($id) {
+        if (!auth()->user()->can('delete_department')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $department = Department::findOrFail($id);
         $department->delete();
         return 'success';
